Change average column type to float Rollback and re-migrate

+++ db/migrate/20181111055151_create_damage_calculations.rb
@@ -3,7 +3,7 @@ class CreateDamageCalculations < ActiveRecord::Migration[5.2]
     create_table :damage_calculations do |t|
       t.integer :dice_count
       t.integer :die_value
-      t.integer :average
+      t.float :average
       t.integer :max
       t.integer :min
       t.integer :user_id
